chore: move extractors into separate package
